Update GrafanaDashboard terminology to ClassicDashboard
This terminology change is made to match our API and UI
experiences to improve consistency for the customer.
