+++ CHANGELOG.md
@@ -9,6 +9,9 @@ Added:
 - Add new resource type, `dataset`, with the first telemetry type supported being tracing.
 - Add new `IN` and `NOT_IN` variations to tracing's `StringFilterMatchType` enum.
 
+Updated:
+- Renamed `chronosphere_grafana_dashboard` to `chronosphere_classic_dashboard`
+
 Deprecated:
 - Block unsupported use of duplicate routes with the same severity in `chronosphere_notification_policy`
 - Remove unsupported `bucket_id` in `chronosphere_notification_policy`
