--- examples/grafana-dashboard/bucket.tf
@@ -1,1 +1,0 @@
-../bucket/main.tf
