--- examples/grafana-dashboard/version.tf
@@ -1,1 +1,0 @@
-../version.tf
